Refactored into new MCTS package
