+++ app/src/main/java/Bamboo/controller/MCTS/MCTS.java
@@ -1,7 +1,8 @@
-package Bamboo.controller;
+package Bamboo.controller.MCTS;
 
+import Bamboo.controller.Agent;
+import Bamboo.controller.Vector;
 import Bamboo.model.Game;
-import Bamboo.model.NodeMCTS;
 
 import java.awt.Color;
